+++ src/OctoshiftCLI.Tests/Octoshift/Commands/ReclaimMannequin/ReclaimMannequinCommandHandlerTests.cs
@@ -12,15 +12,19 @@ public class ReclaimMannequinCommandHandlerTests
 {
     private readonly Mock<OctoLogger> _mockOctoLogger = TestHelpers.CreateMock<OctoLogger>();
     private readonly Mock<ReclaimService> _mockReclaimService = TestHelpers.CreateMock<ReclaimService>();
+    private readonly ConfirmationService _confirmationService;
     private readonly ReclaimMannequinCommandHandler _handler;
 
     private const string GITHUB_ORG = "FooOrg";
     private const string MANNEQUIN_USER = "mona";
     private const string TARGET_USER = "mona_emu";
 
+    private string _consoleOutput;
+
     public ReclaimMannequinCommandHandlerTests()
     {
-        _handler = new ReclaimMannequinCommandHandler(_mockOctoLogger.Object, _mockReclaimService.Object)
+        _confirmationService = new ConfirmationService(CaptureConsoleOutput, MockConsoleKeyPress);
+        _handler = new ReclaimMannequinCommandHandler(_mockOctoLogger.Object, _mockReclaimService.Object, _confirmationService)
         {
             FileExists = _ => true,
             GetFileContent = _ => Array.Empty<string>()
@@ -90,7 +94,7 @@ public class ReclaimMannequinCommandHandlerTests
         };
         await _handler.Handle(args);
 
-        _mockReclaimService.Verify(x => x.ReclaimMannequins(Array.Empty<string>(), GITHUB_ORG, false), Times.Once);
+        _mockReclaimService.Verify(x => x.ReclaimMannequins(Array.Empty<string>(), GITHUB_ORG, false, false), Times.Once);
     }
 
     [Fact]
@@ -105,6 +109,49 @@ public class ReclaimMannequinCommandHandlerTests
         };
         await _handler.Handle(args); // All parameters passed. CSV has precedence
 
-        _mockReclaimService.Verify(x => x.ReclaimMannequins(Array.Empty<string>(), GITHUB_ORG, false), Times.Once);
+        _mockReclaimService.Verify(x => x.ReclaimMannequins(Array.Empty<string>(), GITHUB_ORG, false, false), Times.Once);
     }
+
+    [Fact]
+    public async Task Skip_Invitation_Happy_Path()
+    {
+        // Arrange
+        var expectedResult = "Reclaiming mannequins with the --skip-invitation option is immediate and irreversible. Are you sure you wish to continue? (y/n)Confirmation Recorded. Proceeding...";
+
+        var args = new ReclaimMannequinCommandArgs
+        {
+            GithubOrg = GITHUB_ORG,
+            SkipInvitation = true,
+            Csv = "file.csv",
+        };
+
+        // Act
+        await _handler.Handle(args);
+
+        // Assert
+        _mockReclaimService.Verify(x => x.ReclaimMannequins(Array.Empty<string>(), GITHUB_ORG, false, true), Times.Once);
+        _consoleOutput.Trim().Should().BeEquivalentTo(expectedResult);
+    }
+
+    [Fact]
+    public async Task Skip_Invitation_Without_CSV_Throws_Error()
+    {
+        // Arrange
+        var args = new ReclaimMannequinCommandArgs
+        {
+            GithubOrg = GITHUB_ORG,
+            SkipInvitation = true,
+            MannequinUser = MANNEQUIN_USER,
+            TargetUser = TARGET_USER,
+        };
+
+        // Act/ Assert
+        await FluentActions
+            .Invoking(async () => await _handler.Handle(args))
+            .Should().ThrowAsync<OctoshiftCliException>();
+    }
+
+    private void CaptureConsoleOutput(string msg) => _consoleOutput += msg;
+
+    private ConsoleKey MockConsoleKeyPress() => ConsoleKey.Y;
 }