+++ src/bbs2gh/Program.cs
@@ -43,6 +43,7 @@ namespace OctoshiftCLI.BbsToGithub
                 .AddSingleton<WarningsCountLogger>()
                 .AddSingleton<IVersionProvider, VersionChecker>(sp => sp.GetRequiredService<VersionChecker>())
                 .AddSingleton<BbsArchiveDownloaderFactory>()
+                .AddSingleton<ConfirmationService>()
                 .AddSingleton<ITargetGithubApiFactory>(sp => sp.GetRequiredService<GithubApiFactory>())
                 .AddHttpClient("Kerberos", kerberos: true, noSsl: false)
                 .AddHttpClient("NoSSL", kerberos: false, noSsl: true)