+++ src/gei/Program.cs
@@ -44,6 +44,7 @@ namespace OctoshiftCLI.GithubEnterpriseImporter
                 .AddSingleton<SecretScanningAlertServiceFactory>()
                 .AddSingleton<CodeScanningAlertServiceFactory>()
                 .AddSingleton<GhesVersionCheckerFactory>()
+                .AddSingleton<ConfirmationService>()
                 .AddHttpClient("NoSSL")
                 .ConfigurePrimaryHttpMessageHandler(() => new HttpClientHandler()
                 {